pyproject.toml
[project]
name = "cafe701-monitor"
version = "0.1.0"
requires-python = ">=3.11,<3.14"
dependencies = [
    "flask",
    "pillow",
    "requests",
    "gunicorn",
]
